Switch landing hero components; move docs-specific from components dir to docs-view dir
