+++ src/components/landing-hero/index.tsx
@@ -3,42 +3,37 @@
  * SPDX-License-Identifier: MPL-2.0
  */
 
-import classNames from 'classnames'
-import { useCurrentProduct } from 'contexts'
-import IconTileLogo from 'components/icon-tile-logo'
-import Heading from 'components/heading'
-import Text from 'components/text'
-import { LandingHeroProps } from './types'
+import {
+	CertificationsHero,
+	CertificationsHeroText,
+} from 'views/certifications/components'
+import Image from 'next/image'
+import svgHeroImage from './assets/certification-hero-image.svg'
 import s from './landing-hero.module.css'
 
-const LandingHero = ({ pageHeading, pageSubtitle }: LandingHeroProps) => {
-	const currentProduct = useCurrentProduct()
-
-	const hasSubtitle = !!pageSubtitle
-
+function LandingHero({
+	heading,
+	description,
+}: {
+	heading: string
+	description?: string
+}) {
 	return (
-		<div className={classNames(s.root, { [s.hasSubtitle]: hasSubtitle })}>
-			<IconTileLogo
-				productSlug={
-					currentProduct.slug === 'sentinel' ? 'hcp' : currentProduct.slug
-				}
-				className={s.icon}
-			/>
-			<div>
-				<Heading
-					className={s.pageTitle}
-					id={pageHeading.id}
-					level={1}
-					size={600}
-					weight="bold"
-				>
-					{pageHeading.title}
-				</Heading>
-				{hasSubtitle ? (
-					<Text className={s.pageSubtitle}>{pageSubtitle}</Text>
-				) : null}
-			</div>
-		</div>
+		<CertificationsHero
+			backgroundClassName={s.heroBackground}
+			startSlot={
+				<CertificationsHeroText
+					heading={heading}
+					description={description?.length ? description : ''}
+					foreground="light"
+				/>
+			}
+			endSlot={
+				<div className={s.heroImage}>
+					<Image alt="" src={svgHeroImage} width={447} height={515} />
+				</div>
+			}
+		/>
 	)
 }
