--- src/components/landing-hero/types.ts
@@ -1,11 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: MPL-2.0
- */
-
-import { ProductRootDocsPathLandingProps } from 'views/product-root-docs-path-landing/types'
-
-export interface LandingHeroProps {
-	pageHeading: ProductRootDocsPathLandingProps['pageHeading']
-	pageSubtitle?: string
-}
