+++ src/views/certifications/views/landing/components/index.tsx
@@ -6,4 +6,3 @@
 export * from './certification-program-summary-card'
 export * from './exam-badge-and-title'
 export * from './exam-card'
-export * from './landing-hero'
