--- src/views/certifications/views/landing/components/landing-hero/landing-hero.module.css
@@ -1,38 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: MPL-2.0
- */
-
-.heroBackground {
-	background: url('./assets/certifications-hero-gradient.svg'),
-		var(--token-color-foreground-strong);
-	background-size: cover;
-	background-position: top center;
-	background-repeat: no-repeat;
-
-	@nest html[data-theme='dark'] & {
-		background-color: var(--token-color-page-faint);
-	}
-}
-
-.heroImage {
-	display: flex;
-	height: 200px;
-	overflow: visible;
-	position: relative;
-
-	& > img {
-		position: absolute;
-		top: -40px;
-		left: 50%;
-		transform: translateX(-50%);
-
-		@media only screen and (min-width: 481px) {
-			position: relative;
-			top: -40px;
-			left: auto;
-			margin-left: auto;
-			transform: none;
-		}
-	}
-}
