+++ src/views/product-root-docs-path-landing/components/hero/index.tsx
@@ -6,7 +6,7 @@
 import { useCurrentProduct } from 'contexts'
 import { ProductRootDocsPathLandingIconCardLinkGrid } from 'views/product-root-docs-path-landing/components'
 import DocsPageHeading from 'views/docs-view/components/docs-page-heading'
-import LandingHero from 'components/landing-hero'
+import LandingHero from 'views/docs-view/components/landing-hero'
 import DocsVersionSwitcher from 'components/docs-version-switcher'
 import { ProductRootDocsPathLandingHeroProps } from './types'
 import s from './hero.module.css'
